Bring Kafka version back to 2.6
Now tests with Kafka version 2.6.3. Note that the related Confluent
Platform introduced a dependancy issue with v6.0.3. The tests fail
with a NoSuchMethodError because the Jersey server changed a method
signature in its version 2.34 and that is used by Confluent v6.0.3.
The rest of the system is using Jersey server 2.31.