add support to handle records with null value
tombstone records currently cause the sink-task to stop, as an exception is thrown. Added a new config property "behavior.on.null.values" to control this behavior in a more granular fashion.
The following enum values are allowed:
- ignore (default)
- fail
- log
Also added details (partition,offset) about which record is failing on the given topic.